Moved the overlay files
